musaabi

+++ Cub3d_slippery/libft/ft_substr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yasinsensoy <yasinsensoy@student.42.fr>    +#+  +:+       +#+        */
+/*   By: mtemel <mtemel@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/09 17:34:00 by ysensoy           #+#    #+#             */
-/*   Updated: 2023/02/23 21:01:05 by yasinsensoy      ###   ########.fr       */
+/*   Updated: 2023/03/04 11:30:39 by mtemel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ char	*ft_substr(const char *s, unsigned int start, size_t len)
 	src = (char *)s;
 	if (!src)
 		return (NULL);
-	if (start >= ft_strlen2(s))
+	if (start >= (unsigned int)ft_strlen2(s))
 	{
 		str = (char *)malloc(sizeof(char));
 		if (!str)
